+++ cmd/adjust.go
@@ -1,5 +1,3 @@
-// Copyright (c) 2017 Michael R. Cook
-
 package cmd
 
 import (