+++ tw/main.go
@@ -1,5 +1,4 @@
-// Copyright (c) 2017 Michael R. Cook
-
+// TimeWarrior Copyright (c) 2017 Michael R. Cook
 package main
 
 import (